requirements.txt
requests
openai
python-dotenv
PyMuPDF
google-api-python-client
google-auth
